requirements.txt
torch>=1.12.0
numpy>=1.19.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.7.0
scikit-learn>=1.0.0
tqdm>=4.62.0
pyyaml>=6.0
tensorboard>=2.8.0
transformers>=4.18.0
nltk>=3.6.0
jieba>=0.42.0
safetensors>=0.2.0 